+++ src/main/simplejpa/swing/template/TemplateTableCellRenderer.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013 Jocki Hendry.
+ * Copyright 2015 Jocki Hendry.
  *
  * Licensed under the Apache License, Version 2.0 (the 'License');
  * you may not use this file except in compliance with the License.
@@ -14,41 +14,25 @@
  * limitations under the License.
  */
 
-package simplejpa.swing.glazed.renderer
+package simplejpa.swing.template
 
-import groovy.text.SimpleTemplateEngine
-import groovy.text.Template
-import groovy.text.TemplateEngine
-import simplejpa.swing.TemplateRenderer
 import simplejpa.swing.glazed.ast.ConditionSupport
-
 import javax.swing.table.DefaultTableCellRenderer
 
 @ConditionSupport
 class TemplateTableCellRenderer extends DefaultTableCellRenderer {
 
-    Template template
-    Closure expression
+    TemplateRenderer templateRenderer
 
-    public TemplateTableCellRenderer() {
+    public TemplateTableCellRenderer(def template) {
         super()
-    }
-
-    public TemplateTableCellRenderer(String templateString) {
-        TemplateEngine templateEngine = new SimpleTemplateEngine()
-        template = templateEngine.createTemplate(templateString)
-    }
-
-    public TemplateTableCellRenderer(Closure templateExpression) {
-        this.expression = templateExpression
+        templateRenderer = new TemplateRenderer()
+        templateRenderer.add(template)
     }
 
     @Override
     protected void setValue(Object value) {
-        if (template) {
-            super.setValue(TemplateRenderer.make(template, value))
-        } else if (expression) {
-            super.setValue(TemplateRenderer.make(expression, value))
-        }
+        super.setValue(templateRenderer.make(value))
     }
+
 }